total refactoring

+++ src/main/java/com/example/remindbot/service/impl/callback/DeleteCallbackImpl.java
@@ -1,24 +1,25 @@
 package com.example.remindbot.service.impl.callback;
 
 import static com.example.remindbot.utils.ResponseBuilder.buildResponse;
+import static com.example.remindbot.utils.cash.EventCash.isEventExist;
 
 import com.example.remindbot.model.dto.CallbackWrapper;
 import com.example.remindbot.service.CallbackService;
 import com.example.remindbot.utils.cash.EventCash;
 import com.example.remindbot.utils.cash.StateCash;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import lombok.Setter;
+import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.springframework.stereotype.Component;
 
-@Component
-@Setter
+@Component("delete_reminder_query")
 public class DeleteCallbackImpl implements CallbackService {
 
     @Override
-    public SendMessage handleCallbackQuery(CallbackWrapper wrapper) {
-        EventCash.deleteEventCash(wrapper.getId());
-        StateCash.deleteState(wrapper.getId());
-        return buildResponse(wrapper.getId(), "Removed");
-
+    public BotApiMethod<?> handleCallbackQuery(CallbackWrapper wrapper) {
+        if (isEventExist(wrapper.getId())) {
+            EventCash.deleteEventCash(wrapper.getId());
+            StateCash.deleteState(wrapper.getId());
+            return buildResponse(wrapper.getId(), "Removed");
+        }
+        return buildResponse(wrapper.getId());
     }
 }
